+++ src/Network/Network/VirtualNetworkGateway/VirtualNetworkGatewayBaseCmdlet.cs
@@ -87,5 +87,33 @@ namespace Microsoft.Azure.Commands.Network
             result = result.Replace("virtualNetworkGatewayIpConfiguration", "ipConfigurations");
             return result;
         }
+
+        /// <summary>
+        /// // Backfill Radius server secrets for VPN VirtualNetworkGateway b fetching Radius server secrets using new Post API before calling PUT.
+        /// </summary>
+        /// <param name="psVirtualNetworkGateway">Virtual Network Gateway resource</param>
+        public void BackfillRadiusServerSecrets(PSVirtualNetworkGateway psVirtualNetworkGateway)
+        {
+            if (psVirtualNetworkGateway.VpnClientConfiguration != null && psVirtualNetworkGateway.VpnClientConfiguration.VpnAuthenticationTypes != null && psVirtualNetworkGateway.VpnClientConfiguration.VpnAuthenticationTypes.Contains(VpnAuthenticationType.Radius))
+            {
+                var radiusAuthServers = (List<RadiusAuthServer>)this.VirtualNetworkGatewayClient.ListRadiusSecrets(psVirtualNetworkGateway.ResourceGroupName, psVirtualNetworkGateway.Name).Value;
+
+                if (radiusAuthServers != null && radiusAuthServers.Any())
+                {
+                    if (!string.IsNullOrWhiteSpace(psVirtualNetworkGateway.VpnClientConfiguration.RadiusServerAddress) && string.IsNullOrWhiteSpace(psVirtualNetworkGateway.VpnClientConfiguration.RadiusServerSecret))
+                    {
+                        psVirtualNetworkGateway.VpnClientConfiguration.RadiusServerSecret = radiusAuthServers.Find(radius => radius.RadiusServerAddress == psVirtualNetworkGateway.VpnClientConfiguration.RadiusServerAddress).RadiusServerSecret ?? "";
+                    }
+
+                    if (psVirtualNetworkGateway.VpnClientConfiguration.RadiusServers != null && psVirtualNetworkGateway.VpnClientConfiguration.RadiusServers.Any())
+                    {
+                        foreach (var radiusServer in psVirtualNetworkGateway.VpnClientConfiguration.RadiusServers)
+                        {
+                            radiusServer.RadiusServerSecret = radiusAuthServers.Find(radius => radius.RadiusServerAddress == radiusServer.RadiusServerAddress).RadiusServerSecret ?? "";
+                        }
+                    }
+                }
+            }
+        }
     }
 }